requirements.txt
astroid==2.5
isort==5.7.0
lazy-object-proxy==1.5.2
mccabe==0.6.1
paho-mqtt==1.5.1
pkg-resources==0.0.0
pylint==2.7.1
toml==0.10.2
typed-ast==1.4.2
vcgencmd==0.1.1
wrapt==1.12.1
